Implemented executeScript() method
Do My Own Pest Control is considering using DreamFactory as the backend for our mobile app that we are starting development on.  We will need a fair number of pre and post processing scripts that we will write in PHP and we want to be able to manage those scripts using git and BitBucket.  By implementing the executeScript() method, we can have our own Laravel project that we will use to contain those scripts, named for the event, like dmopc._table.categories.get.post_process.php.  We will add an artisan command that will iterate over this directory, and using the basename of the file, check if there is a db entry in event_script for that event.  If there is not, it will add one, using the full path to the file for the value of content.